feat: :sparkles: Protected school deletion

+++ apps/api/src/school/domain/services/school/delete_school.service.ts
@@ -2,16 +2,24 @@ import { inject } from '@adonisjs/core'
 import { ISchoolsRepository } from '#school/domain/contracts/repositories/schools.repository'
 import { Id } from '#shared/id/domain/models/id'
 import { SchoolNotFoundException } from '#school/domain/models/school_not_found.exception'
+import { UnauthorizedException } from '#shared/domain/exceptions/unauthorized.exception'
+import { User } from '#user/domain/models/user'
 
 @inject()
 export class DeleteSchoolService {
   constructor(private readonly schoolsRepository: ISchoolsRepository) {}
 
-  async execute(schoolId: Id): Promise<void> {
+  async execute(schoolId: Id, user: User): Promise<void> {
     const school = await this.schoolsRepository.getById(schoolId)
 
     if (!school) {
       throw new SchoolNotFoundException(schoolId)
+    }
+
+    const admins = await this.schoolsRepository.getAdmins(schoolId)
+
+    if (!admins.find((admin) => admin.id.equals(user.id))) {
+      throw new UnauthorizedException()
     }
 
     await this.schoolsRepository.deleteById(schoolId)
